# contracts/openapi.yaml
components:
  schemas:
    AcceptedJob:
      additionalProperties: false
      properties:
        job_id:
          pattern: ^[0-9A-HJKMNP-TV-Z]{26}$
          title: Job Id
          type: string
        retry_after_seconds:
          default: 2
          maximum: 60.0
          minimum: 1.0
          title: Retry After Seconds
          type: integer
        status_url:
          title: Status Url
          type: string
      required:
      - job_id
      - status_url
      title: AcceptedJob
      type: object
...
    AppStoreVerificationRequest:
      additionalProperties: false
      properties:
        signed_transaction:
          format: password
          maxLength: 32768
          minLength: 10
          title: Signed Transaction
          type: string
          writeOnly: true
        store:
          const: app_store
          title: Store
          type: string
      required:
      - store
      - signed_transaction
      title: AppStoreVerificationRequest
      type: object
    BacktestEvaluation:
      additionalProperties: false
      properties:
        metrics:
          $ref: '#/components/schemas/BacktestMetrics'
        model_name:
          maxLength: 64
          minLength: 1
          title: Model Name
          type: string
        model_version:
          maxLength: 128
          minLength: 1
          title: Model Version
          type: string
        residuals:
          items:
            type: number
          maxLength: 100000
          minLength: 1
          title: Residuals
          type: array
      required:
      - model_name
      - model_version
      - metrics
      - residuals
      title: BacktestEvaluation
      type: object
    BacktestMetrics:
      additionalProperties: false
      properties:
        interval_coverage:
          anyOf:
          - maximum: 1.0
            minimum: 0.0
            type: number
          - type: 'null'
          title: Interval Coverage
        mae:
          minimum: 0.0
          title: Mae
          type: number
        mase:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
          title: Mase
        wape:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
          title: Wape
      required:
      - mae
      title: BacktestMetrics
      type: object
    Body_receipt_job_create:
      properties:
        file:
          contentMediaType: application/octet-stream
          title: File
          type: string
      required:
      - file
      title: Body_receipt_job_create
      type: object
    CoachAnswer:
      additionalProperties: false
      properties:
        answer:
          anyOf:
          - maxLength: 4000
            minLength: 1
            type: string
          - type: 'null'
          title: Answer
        safety:
          $ref: '#/components/schemas/CoachSafety'
        sources:
          items:
            $ref: '#/components/schemas/SourceCard'
          title: Sources
          type: array
      required:
      - safety
      - sources
      title: CoachAnswer
      type: object
    CoachJobResult:
      additionalProperties: false
      properties:
        answer:
          $ref: '#/components/schemas/CoachAnswer'
        kind:
          const: coach
          title: Kind
          type: string
        schema_version:
          const: 1
          title: Schema Version
          type: integer
      required:
      - kind
      - schema_version
      - answer
      title: CoachJobResult
      type: object
    CoachQuery:
      additionalProperties: false
      properties:
        locale:
          const: tr-TR
          default: tr-TR
          title: Locale
          type: string
        question:
          maxLength: 2000
          minLength: 1
          title: Question
          type: string
        session_id:
          maxLength: 26
          minLength: 26
          pattern: ^[0-9A-HJKMNP-TV-Z]{26}$
          title: Session Id
          type: string
      required:
      - question
      - session_id
      title: CoachQuery
      type: object
    CoachSafety:
      enum:
      - answered
      - insufficient_sources
      title: CoachSafety
      type: string
...
    DataExport:
      additionalProperties: false
      properties:
        billing_events:
          items:
            $ref: '#/components/schemas/ExportBillingEvent'
          title: Billing Events
          type: array
        entitlements:
          items:
            $ref: '#/components/schemas/ExportEntitlement'
          title: Entitlements
          type: array
        generated_at:
          format: date-time
          title: Generated At
          type: string
        jobs:
          items:
            $ref: '#/components/schemas/ExportJob'
          title: Jobs
          type: array
        schema_version:
          const: '1.0'
          default: '1.0'
          title: Schema Version
          type: string
      required:
      - generated_at
      - entitlements
      - jobs
      - billing_events
      title: DataExport
      type: object
    DeletionCounts:
      additionalProperties: false
      properties:
        billing_events_anonymized:
          default: 0
          minimum: 0.0
          title: Billing Events Anonymized
          type: integer
        entitlements:
          default: 0
          minimum: 0.0
          title: Entitlements
          type: integer
        jobs:
          default: 0
          minimum: 0.0
          title: Jobs
          type: integer
        leaderboard_contributions:
          default: 0
          minimum: 0.0
          title: Leaderboard Contributions
          type: integer
      title: DeletionCounts
      type: object
...
    EntitlementCollection:
      additionalProperties: false
      properties:
        items:
          items:
            $ref: '#/components/schemas/EntitlementView'
          title: Items
          type: array
      required:
      - items
      title: EntitlementCollection
      type: object
    EntitlementStatus:
      enum:
      - active
      - grace_period
      - expired
      - revoked
      - refunded
      title: EntitlementStatus
      type: string
    EntitlementView:
      additionalProperties: false
      properties:
        expires_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Expires At
        product_id:
          title: Product Id
          type: string
        status:
          $ref: '#/components/schemas/EntitlementStatus'
        store:
          $ref: '#/components/schemas/Store'
      required:
      - store
      - product_id
      - status
      - expires_at
      title: EntitlementView
      type: object
    ExportBillingEvent:
      additionalProperties: false
      properties:
        event:
          $ref: '#/components/schemas/StoreEvent'
        expires_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Expires At
        occurred_at:
          format: date-time
          title: Occurred At
          type: string
        product_id:
          pattern: ^[a-z0-9_.-]{3,128}$
          title: Product Id
          type: string
        store:
          $ref: '#/components/schemas/Store'
      required:
      - store
      - product_id
      - event
      - occurred_at
      - expires_at
      title: ExportBillingEvent
      type: object
    ExportEntitlement:
      additionalProperties: false
      properties:
        expires_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Expires At
        product_id:
          pattern: ^[a-z0-9_.-]{3,128}$
          title: Product Id
          type: string
        status:
          $ref: '#/components/schemas/EntitlementStatus'
        store:
          $ref: '#/components/schemas/Store'
      required:
      - store
      - product_id
      - status
      - expires_at
      title: ExportEntitlement
      type: object
    ExportJob:
      additionalProperties: false
      properties:
        completed_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Completed At
        created_at:
          format: date-time
          title: Created At
          type: string
        job_id:
          pattern: ^[0-9A-HJKMNP-TV-Z]{26}$
          title: Job Id
          type: string
        kind:
          $ref: '#/components/schemas/JobKind'
        status:
          $ref: '#/components/schemas/JobStatus'
      required:
      - job_id
      - kind
      - status
      - created_at
      - completed_at
      title: ExportJob
      type: object
    FieldConfidence:
      additionalProperties: false
      properties:
        confidence:
          maximum: 1.0
          minimum: 0.0
          title: Confidence
          type: number
        field_name:
          maxLength: 64
          minLength: 1
          title: Field Name
          type: string
      required:
      - field_name
      - confidence
      title: FieldConfidence
      type: object
    ForecastCandidate:
      additionalProperties: false
      properties:
        model_name:
          maxLength: 64
          minLength: 1
          title: Model Name
          type: string
        model_version:
          maxLength: 128
          minLength: 1
          title: Model Version
          type: string
        points:
          items:
            $ref: '#/components/schemas/ForecastPoint'
          maxLength: 365
          minLength: 1
          title: Points
          type: array
      required:
      - model_name
      - model_version
      - points
      title: ForecastCandidate
      type: object
...
    ForecastJobResult:
      additionalProperties: false
      properties:
        forecast:
          $ref: '#/components/schemas/ForecastResult'
        kind:
          const: forecast
          title: Kind
          type: string
        schema_version:
          const: 1
          title: Schema Version
          type: integer
      required:
      - kind
      - schema_version
      - forecast
      title: ForecastJobResult
      type: object
    ForecastPoint:
      additionalProperties: false
      properties:
        horizon_day:
          minimum: 1.0
          title: Horizon Day
          type: integer
        lower:
          anyOf:
          - type: number
          - type: 'null'
          title: Lower
        prediction:
          title: Prediction
          type: number
        upper:
          anyOf:
          - type: number
          - type: 'null'
          title: Upper
      required:
      - horizon_day
      - prediction
      title: ForecastPoint
      type: object
    ForecastResult:
      additionalProperties: false
      properties:
        cache_hit:
          title: Cache Hit
          type: boolean
        evaluation:
          $ref: '#/components/schemas/BacktestEvaluation'
        forecast:
          $ref: '#/components/schemas/ForecastCandidate'
        selection_reason:
          $ref: '#/components/schemas/SelectionReason'
      required:
      - forecast
      - evaluation
      - selection_reason
      - cache_hit
      title: ForecastResult
      type: object
    GooglePlayVerificationRequest:
      additionalProperties: false
      properties:
        package_name:
          maxLength: 256
          minLength: 3
          pattern: ^[A-Za-z][A-Za-z0-9_]*(?:\.[A-Za-z0-9_]+)+$
          title: Package Name
          type: string
        purchase_token:
          format: password
          maxLength: 4096
          minLength: 1
          title: Purchase Token
          type: string
          writeOnly: true
        store:
          const: google_play
          title: Store
          type: string
      required:
      - store
      - package_name
      - purchase_token
      title: GooglePlayVerificationRequest
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    JobKind:
      enum:
      - coach
      - forecast
      - receipt
      - model_training
      - billing_verification
      - retention
      title: JobKind
      type: string
    JobStatus:
      enum:
      - accepted
      - queued
      - running
      - retry_wait
      - succeeded
      - failed
      title: JobStatus
      type: string
    JobStatusView:
      additionalProperties: false
      properties:
        failure_code:
          anyOf:
          - maxLength: 64
            type: string
          - type: 'null'
          title: Failure Code
        job_id:
          pattern: ^[0-9A-HJKMNP-TV-Z]{26}$
          title: Job Id
          type: string
        kind:
          $ref: '#/components/schemas/JobKind'
        result:
          anyOf:
          - discriminator:
              mapping:
                coach: '#/components/schemas/CoachJobResult'
                forecast: '#/components/schemas/ForecastJobResult'
                receipt: '#/components/schemas/ReceiptJobResult'
              propertyName: kind
            oneOf:
            - $ref: '#/components/schemas/CoachJobResult'
            - $ref: '#/components/schemas/ReceiptJobResult'
            - $ref: '#/components/schemas/ForecastJobResult'
          - type: 'null'
          title: Result
        result_state:
          $ref: '#/components/schemas/ResultState'
        status:
          $ref: '#/components/schemas/JobStatus'
        updated_at:
          format: date-time
          title: Updated At
          type: string
      required:
      - job_id
      - kind
      - status
      - result_state
      - updated_at
      title: JobStatusView
      type: object
...
    ReceiptItem:
      additionalProperties: false
      properties:
        confidence:
          maximum: 1.0
          minimum: 0.0
          title: Confidence
          type: number
        line_total_minor:
          minimum: 0.0
          title: Line Total Minor
          type: integer
        product_name:
          maxLength: 200
          minLength: 1
          title: Product Name
          type: string
        quantity:
          decimal_places: 18
          max_digits: 38
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Quantity
          type: string
        unit_price_minor:
          minimum: 0.0
          title: Unit Price Minor
          type: integer
      required:
      - product_name
      - quantity
      - unit_price_minor
      - line_total_minor
      - confidence
      title: ReceiptItem
      type: object
    ReceiptJobResult:
      additionalProperties: false
      properties:
        kind:
          const: receipt
          title: Kind
          type: string
        receipt:
          $ref: '#/components/schemas/ReceiptResult'
        schema_version:
          const: 1
          title: Schema Version
          type: integer
      required:
      - kind
      - schema_version
      - receipt
      title: ReceiptJobResult
      type: object
    ReceiptResult:
      additionalProperties: false
      properties:
        discount_minor:
          default: 0
          minimum: 0.0
          title: Discount Minor
          type: integer
        field_confidences:
          items:
            $ref: '#/components/schemas/FieldConfidence'
          maxLength: 600
          title: Field Confidences
          type: array
        items:
          items:
            $ref: '#/components/schemas/ReceiptItem'
          maxLength: 500
          title: Items
          type: array
        merchant_name:
          anyOf:
          - maxLength: 200
            minLength: 1
            type: string
          - type: 'null'
          title: Merchant Name
        requires_review:
          title: Requires Review
          type: boolean
        subtotal_minor:
          anyOf:
          - minimum: 0.0
            type: integer
          - type: 'null'
          title: Subtotal Minor
        tax_minor:
          default: 0
          minimum: 0.0
          title: Tax Minor
          type: integer
        total_minor:
          anyOf:
          - minimum: 0.0
            type: integer
          - type: 'null'
          title: Total Minor
      required:
      - items
      - field_confidences
      - requires_review
      title: ReceiptResult
      type: object
...
    ResultState:
      enum:
      - pending
      - ready
      - expired
      - failed
      title: ResultState
      type: string
    SelectionReason:
      enum:
      - baseline_wape
      - candidate_coverage
      - candidate_improvement
      - candidate_error
      - baseline_only
      title: SelectionReason
      type: string
    SourceCard:
      additionalProperties: false
      properties:
        institution:
          maxLength: 64
          minLength: 2
          title: Institution
          type: string
        period:
          format: date
          title: Period
          type: string
        release_date:
          format: date
          title: Release Date
          type: string
        retrieved_at:
          format: date-time
          title: Retrieved At
          type: string
        series_id:
          pattern: ^[A-Z0-9_.-]{2,64}$
          title: Series Id
          type: string
        snapshot_digest:
          pattern: ^[0-9a-f]{64}$
          title: Snapshot Digest
          type: string
        snapshot_id:
          maxLength: 26
          minLength: 26
          pattern: ^[0-9A-HJKMNP-TV-Z]{26}$
          title: Snapshot Id
          type: string
        source_url:
          format: uri
          maxLength: 2083
          minLength: 1
          title: Source Url
          type: string
        unit:
          maxLength: 32
          minLength: 1
          title: Unit
          type: string
        value:
          decimal_places: 18
          max_digits: 38
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Value
          type: string
      required:
      - snapshot_id
      - snapshot_digest
      - institution
      - series_id
      - period
      - value
      - unit
      - release_date
      - source_url
      - retrieved_at
      title: SourceCard
      type: object
    Store:
      enum:
      - google_play
      - app_store
      title: Store
      type: string
    StoreEvent:
      enum:
      - purchased
      - renewed
      - grace_period
      - canceled
      - expired
      - revoked
      - refunded
      - restored
      title: StoreEvent
      type: string
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  description: Maki mobil finans uygulamasının güvenli servis sınırı.
  title: Maki API
  version: 1.0.0
openapi: 3.1.0
paths:
  /api/v1/billing/entitlements:
    get:
      description: Oturum sahibinin doğrulanmış abonelik haklarını döndürür.
      operationId: billing_entitlement_list
      parameters:
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntitlementCollection'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Entitlements
      tags:
      - abonelik
  /api/v1/billing/verifications:
    post:
      description: Mağaza kanıtını sunucuda doğrular ve abonelik hakkını günceller.
      operationId: billing_verification_create
      parameters:
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          maxLength: 128
          minLength: 8
          title: Idempotency-Key
          type: string
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              discriminator:
                mapping:
                  app_store: '#/components/schemas/AppStoreVerificationRequest'
                  google_play: '#/components/schemas/GooglePlayVerificationRequest'
                propertyName: store
              oneOf:
              - $ref: '#/components/schemas/GooglePlayVerificationRequest'
              - $ref: '#/components/schemas/AppStoreVerificationRequest'
              title: Payload
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntitlementView'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Verification
      tags:
      - abonelik
  /api/v1/coach/queries:
    post:
      description: Temizlenmiş tek koç sorusunu iş kuyruğuna kabul eder.
      operationId: coach_query_create
      parameters:
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          maxLength: 128
          minLength: 8
          title: Idempotency-Key
          type: string
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoachQuery'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedJob'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Coach Query
      tags:
      - koç
...
  /api/v1/jobs/{job_id}:
    get:
      description: Yalnız oturum sahibine ait iş durumunu döndürür.
      operationId: job_status_get
      parameters:
      - in: path
        name: job_id
        required: true
        schema:
          pattern: ^[0-9A-HJKMNP-TV-Z]{26}$
          title: Job Id
          type: string
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatusView'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Job Status
      tags:
      - işler
...
  /api/v1/privacy/data:
    delete:
      description: Oturum sahibine bağlı kayıtları siler veya geri bağlanamaz hâle getirir.
      operationId: privacy_data_delete
      parameters:
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          maxLength: 128
          minLength: 8
          title: Idempotency-Key
          type: string
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletionCounts'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Data
      tags:
      - gizlilik
  /api/v1/privacy/exports:
    get:
      description: Oturum sahibine bağlı sunucu kayıtlarını güvenli biçimde dışa aktarır.
      operationId: privacy_export_get
      parameters:
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataExport'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Export Data
      tags:
      - gizlilik
  /api/v1/receipts/jobs:
    post:
      description: Fiş görselini kısa ömürlü OCR kuyruğuna kabul eder.
      operationId: receipt_job_create
      parameters:
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          maxLength: 128
          minLength: 8
          title: Idempotency-Key
          type: string
      - in: header
        name: Authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_receipt_job_create'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedJob'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Receipt Job
      tags:
      - fiş
...